requirements.txt
flask==2.3.2
flask-socketio==5.3.4
python-dotenv==1.0.0
ccxt==3.1.30
pandas==2.0.3
numpy==1.24.3
ta==0.11.0  # Alternative to TA-Lib that's easier to install
python-engineio==4.5.1
python-socketio==5.8.0

